+++ hobby.csv
@@ -1,0 +1,2 @@
+скалолазание,охота
+горные лыжи
