+++ users.csv
@@ -1,0 +1,3 @@
+Иванов,Иван,Иванович
+Петров,Петр,Петрович
+Егоров,Егор,Егорович
