+++ users_hobby.txt
@@ -1,0 +1,1 @@
+{"Иванов,Иван,Иванович": "скалолазание,охота", "Петров,Петр,Петрович": "горные лыжи", "Егоров,Егор,Егорович": null}
